+++ sdmx-dl-cli/src/test/java/sdmxdl/cli/ListSourcesCommandTest.java
@@ -41,7 +41,7 @@ public class ListSourcesCommandTest {
         assertThat(watcher.getOut())
                 .isEmpty();
         assertThat(watcher.getErrWithoutAnsiCodes())
-                .contains("[CFG] RI_REGISTRY: Using 1 custom sources from file " + Paths.get(src.getPath()).toUri())
+                .contains("[-] (RI_REGISTRY) Using 1 custom sources from file " + Paths.get(src.getPath()).toUri())
                 .doesNotContain("[SSL] Initializing SSL factory");
 
         assertThat(FileSample.readAll(out))