+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+
+- ![CLI] Improve CLI message and error reporting [#1243](https://github.com/nbbrd/sdmx-dl/issues/1243)
+
 ### Fixed
 
 - ![PROVIDER] Fix UIS SDMX API end-of-life [#901](https://github.com/nbbrd/sdmx-dl/issues/901)